exclude some non-pid values
git-svn-id: https://svn.r-project.org/R/trunk@56691 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/tools/src/signals.c
@@ -54,6 +54,8 @@ SEXP ps_kill(SEXP spid, SEXP ssignal)
                 CloseHandle(hProc);
             }
 #elif defined(HAVE_KILL)
+            if (pid[i] <= 0)
+                continue;
             if (pid[i] != NA_INTEGER)
                 res[i] = kill(pid[i], signal);
 #endif
@@ -80,6 +82,8 @@ SEXP ps_priority(SEXP spid, SEXP svalue)
     res = INTEGER(sres);
     for (int i = 0; i < ns; i++)
     {
+        if (pid[i] <= 0)
+            continue;
         if (pid[i] != NA_INTEGER)
         {
             /* return value -1 is both an error value
